+++ packages/comment-widget/src/avatar/providers/gravatar.ts
@@ -2,11 +2,11 @@ import AvatarProvider from "./avatar-provider";
 
 class Gravatar extends AvatarProvider {
   override getAvatarSrc(emailHash: string | undefined): string {
-    return this.url + emailHash;
+    return `${this.url}/avatar/${emailHash}`;
   }
 }
 
 export default new Gravatar(
   "Gravatar",
-  "https://gravatar.com/avatar/"
+  "https://gravatar.com"
 );
