+++ packages/comment-widget/src/comment-widget.ts
@@ -27,7 +27,7 @@ import './comment-pagination';
 import varStyles from './styles/var';
 import { ToastManager } from './lit-toast';
 import { AnonymousUserPolicy, AllUserPolicy, NoAvatarUserPolicy, AvatarPolicyEnum, setPolicyInstance } from './avatar/avatar-policy';
-import { setAvatarProvider } from "./avatar-provider";
+import { setAvatarProvider } from "./avatar/providers";
 
 export class CommentWidget extends LitElement {
   @provide({ context: baseUrlContext })
